# docs/swagger.yaml
basePath: /
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  internal_rest_api_user.GetUserByIDOut:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      email:
        type: string
      reputation_score:
        type: integer
      role:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  internal_rest_api_user.GetUserMeOut:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      email:
        type: string
      reputation_score:
        type: integer
      role:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  internal_rest_api_user.PatchUpdateAvatarOut:
    properties:
      avatar_url:
        type: string
    type: object
  internal_rest_api_user.PatchUpdateUserIn:
    properties:
      bio:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  internal_rest_api_user.PatchUpdateUserOut:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      email:
        type: string
      reputation_score:
        type: integer
      role:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: lyryx API
  version: "1.0"
paths:
...
  /v1/user/{id}:
    get:
      description: Возвращает полную информацию о профиле пользователя по его id.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ с профилем пользователя
          schema:
            $ref: '#/definitions/internal_rest_api_user.GetUserByIDOut'
        "400":
          description: Некорректный id пользователя
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Получение данных пользователя по его id.
      tags:
      - user
  /v1/user/me:
    get:
      description: Возвращает полную информацию о профиле текущего пользователя, идентифицированного
        по access_token.
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ с профилем пользователя
          schema:
            $ref: '#/definitions/internal_rest_api_user.GetUserMeOut'
        "401":
          description: Пользователь не аутентифицирован
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Получение данных своего профиля.
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Обновляет email, username, bio или password текущего пользователя.
        Достаточно передать только нужные поля.
      parameters:
      - description: Параметры обновления профиля
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_rest_api_user.PatchUpdateUserIn'
      produces:
      - application/json
      responses:
        "200":
          description: Профиль пользователя обновлен
          schema:
            $ref: '#/definitions/internal_rest_api_user.PatchUpdateUserOut'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Пользователь не аутентифицирован
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "409":
          description: Email или username уже заняты
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Частичное обновление профиля пользователя
      tags:
      - user
  /v1/user/me/avatar:
    patch:
      consumes:
      - multipart/form-data
      description: Принимает avatar в multipart/form-data, валидирует изображение,
        загружает в MinIO и обновляет ссылку на аватар текущего пользователя.
      parameters:
      - description: Файл аватарки (png/jpeg)
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Аватар пользователя обновлен
          schema:
            $ref: '#/definitions/internal_rest_api_user.PatchUpdateAvatarOut'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Пользователь не аутентифицирован
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Обновление аватарки пользователя
      tags:
      - user
swagger: "2.0"